--- lib/app/ListenScreen.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ListenScreen extends StatelessWidget {
-  const ListenScreen({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Expanded(
-        child: Container(
-          color: Colors.white,
-        ),
-      ),
-    );
-  }
-}
