+++ pubspec.yaml
@@ -36,7 +36,6 @@ dependencies:
   cupertino_icons: ^1.0.2
   audioplayers: ^0.20.1
   flutter_native_splash: ^2.1.1
-  google_fonts: ^2.3.2
 
 
 dev_dependencies:
